Check for text in title
In some cases we are failing to find the 'Google' text on the
page. This appears to be because the text 'Google' is actually not
present on the page, rather than an issue with the TestObject platform.

To improve the robustness of the test, we check for the text 'Google' in
the title instead.